Change disk config, adds a map in between

+++ flink-connector-bigquery-integration-test/src/main/java/com/google/cloud/flink/bigquery/integration/BigQueryIntegrationTest.java
@@ -311,7 +311,7 @@ public class BigQueryIntegrationTest {
                                     public GenericRecord map(GenericRecord genericRecord)
                                             throws Exception {
                                         genericRecord.put(
-                                                "number", (int) genericRecord.get("number") + 1);
+                                                "number", (long) genericRecord.get("number") + 1);
                                         return genericRecord;
                                     }
                                 })
